:art: Move router in core folder #2
