+++ src/core/Router.php
@@ -1,5 +1,5 @@
 <?php
-namespace App;
+namespace App\core;
 
 use AltoRouter;
